store/urls.py
# from django.urls import path, include
from . import views
from rest_framework_nested import routers

router = routers.DefaultRouter()

router.register('products', views.ProductViewSet, basename='product')
router.register('categories', views.CategoryViewSet, basename='category')

product_router = routers.NestedDefaultRouter(
	router,
	'products',
	lookup='product',
)
product_router.register('comments', views.CommentViewSet, basename='product-comments')

urlpatterns = router.urls + product_router.urls
