requirements.txt
asgiref==3.7.2
Django==5.0
django-debug-toolbar==4.2.0
django-filter==23.5
django-stubs==4.2.7
django-stubs-ext==4.2.7
djangorestframework==3.14.0
drf-nested-routers==0.93.5
factory-boy==3.3.0
Faker==21.0.0
psycopg2-binary==2.9.9
python-dateutil==2.8.2
pytz==2023.3.post1
six==1.16.0
sqlparse==0.4.4
types-pytz==2023.3.1.1
types-PyYAML==6.0.12.12
typing_extensions==4.9.0
